+++ lib/srv/app/server_test.go
@@ -279,6 +279,13 @@ func SetUpSuiteWithConfig(t *testing.T, config suiteConfig) *Suite {
 	require.NoError(t, err)
 	err = s.appServer.ForceHeartbeat()
 	require.NoError(t, err)
+	t.Cleanup(func() {
+		s.appServer.Close()
+
+		// wait for the server to close before allowing other cleanup
+		// actions to proceed
+		s.appServer.Wait()
+	})
 
 	return s
 }